Fixed issue with QSBenefitProvider deserialization, updated changelog, bump snapshot version.

+++ quickshop-bukkit/src/main/java/com/ghostchu/quickshop/economy/QSBenefitProvider.java
@@ -68,7 +68,7 @@ public class QSBenefitProvider implements BenefitProvider {
     if(CommonUtil.isEmptyString(json)) {
       return new QSBenefitProvider();
     }
-    final Map<String, BigDecimal> map = JsonUtil.regular().fromJson(json, new TypeToken<Map<String, Double>>() {
+    final Map<String, BigDecimal> map = JsonUtil.regular().fromJson(json, new TypeToken<Map<String, BigDecimal>>() {
     }.getType());
 
     final Map<QUser, BigDecimal> parsed = new HashMap<>();
